requirements.txt
openai
pandas
python-dotenv
re
json
os
ast
glob
